block: move block logic out from func
